+++ index.js
@@ -1,3 +1,6 @@
 // This is where your project starts.
 
+const { start } = require('./src/game');
+
 console.log('Your project is running...'); 
+start();
