+++ src/deck.js
@@ -1,0 +1,12 @@
+const createDeck = (deck) => {
+  return deck;
+};
+
+const countCards = (deck) => {
+  return deck.length;
+}
+
+module.exports = {
+  createDeck,
+  countCards
+}
