+++ src/turns.js
@@ -1,0 +1,10 @@
+const evaluateGuess = (guess, card) => {
+  if (guess === card.correctAnswer) {
+    return 'correct!';
+  }
+  return 'incorrect!';
+};
+
+module.exports = {
+  evaluateGuess
+}
